SP-619 Add additional tests Move simple to main code base
