+++ src/main/java/uk/gov/justice/digital/oasys/controller/OffenderController.java
@@ -8,9 +8,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import uk.gov.justice.digital.oasys.api.simple.OffenderDto;
+import uk.gov.justice.digital.oasys.api.OffenderDto;
 import uk.gov.justice.digital.oasys.service.OffenderService;
-import uk.gov.justice.digital.oasys.utils.LogEvent;
 
 @RestController
 @Api(tags = "Offenders")
