+++ src/main/java/uk/gov/justice/digital/oasys/service/AssessmentsService.java
@@ -4,8 +4,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import uk.gov.justice.digital.oasys.api.simple.AssessmentDto;
-import uk.gov.justice.digital.oasys.api.simple.AssessmentSummaryDto;
+import uk.gov.justice.digital.oasys.api.AssessmentDto;
+import uk.gov.justice.digital.oasys.api.AssessmentSummaryDto;
 import uk.gov.justice.digital.oasys.jpa.entity.RefSection;
 import uk.gov.justice.digital.oasys.jpa.entity.simple.Assessment;
 import uk.gov.justice.digital.oasys.jpa.entity.simple.Section;
