+++ src/test/java/uk/gov/justice/digital/oasys/api/ObjectiveDtoTest.java
@@ -2,9 +2,7 @@ package uk.gov.justice.digital.oasys.api;
 
 import org.junit.Test;
 import uk.gov.justice.digital.oasys.controller.ControllerServiceTestContext;
-
 import java.time.LocalDateTime;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class ObjectiveDtoTest {
